renaming registrationserver to server
